requirements.txt
aiosqlite>=0.19.0
aiomqtt>=2.0.0
pyyaml>=6.0
psutil>=5.9.0
